Deleted scrap, fixed collision
Deleted Player Dish and HUD scripts, sprite, etc. Need to understand how to implement before adding in to lessen junk. Fixed collisions, "is trigger" needs to be on in Greedy Students for "take bite" animations to work, allows for Greedy Student to phase through player (which is not desired ofc). Need to fix that, probably with a bool.